worker/requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
schema==0.7.5
